Preference関連の分離
