+++ app/src/main/java/com/example/kotlintodo/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.kotlintodo.model.EditorMode
-import com.example.kotlintodo.common.toDoAccessor
+import com.example.kotlintodo.realm.ToDoAccessor
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 
 class MainActivity : AppCompatActivity() {
@@ -21,34 +21,34 @@ class MainActivity : AppCompatActivity() {
         val addButton = findViewById<FloatingActionButton>(R.id.addButton)
 
 //         仮データ作成
-//         for (i in 1..20) toDoAccessor.create("test${i}")
+//         for (i in 1..20) ToDoAccessor.create("test${i}")
 
 //         全件取得
-        val todos = toDoAccessor.getAll()
+        val todos = ToDoAccessor.getAll()
         Log.d("todos", todos.toString())
 
 //         id=1 更新
-//        toDoAccessor.update(1, "update test")
+//        ToDoAccessor.update(1, "update test")
 //        Log.d("updated todo", ToDoAccessor.find(1).toString())
 
 //         id=20 削除
-//         toDoAccessor.delete(20)
-//         Log.d("todos after deleted 20", toDoAccessor.getAll().toString())
+//         ToDoAccessor.delete(20)
+//         Log.d("todos after deleted 20", ToDoAccessor.getAll().toString())
 
 //         存在しない id=21 更新
-//         toDoAccessor.update(21, "no data update test")
-//         Log.d("todos after no data updated", toDoAccessor.getAll().toString())
+//         ToDoAccessor.update(21, "no data update test")
+//         Log.d("todos after no data updated", ToDoAccessor.getAll().toString())
 
 //         存在しない id=21 削除
-//         toDoAccessor.delete(21)
-//         Log.d("todos after no data updated", toDoAccessor.getAll().toString())
+//         ToDoAccessor.delete(21)
+//         Log.d("todos after no data updated", ToDoAccessor.getAll().toString())
 
 //         存在しない id=21 取得
-//         toDoAccessor.find(21)
+//         ToDoAccessor.find(21)
 
 //         存在する id=1 登録
 //         テスト前にToDoAccessor.createを要修正
-//        toDoAccessor.create("create existed id data")
+//        ToDoAccessor.create("create existed id data")
 
 
         val separateLine = DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
