--- app/src/main/java/com/example/kotlintodo/common/config.kt
@@ -1,5 +1,0 @@
-package com.example.kotlintodo.common
-
-import com.example.kotlintodo.realm.ToDoAccessor
-
-val toDoAccessor = ToDoAccessor
